organised ticket files to a folder and added social-link api and model
