+++ app/src/main/java/org/fossasia/openevent/general/event/EventDetailsFragment.kt
@@ -12,6 +12,7 @@ import com.squareup.picasso.Picasso
 import kotlinx.android.synthetic.main.content_event.view.*
 import org.fossasia.openevent.general.MainActivity
 import org.fossasia.openevent.general.R
+import org.fossasia.openevent.general.ticket.TicketsFragment
 import org.koin.android.architecture.ext.viewModel
 import timber.log.Timber
