+++ app/src/main/java/org/fossasia/openevent/general/social/SocialLinkApi.kt
@@ -1,0 +1,12 @@
+package org.fossasia.openevent.general.social
+
+import io.reactivex.Flowable
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface SocialLinkApi {
+
+    @GET("events/{id}/social-links")
+    fun getSocialLinks(@Path("id") id: Long): Flowable<List<SocialLink>>
+
+}
