+++ app/src/main/java/org/fossasia/openevent/general/ticket/TicketsFragment.kt
@@ -1,4 +1,4 @@
-package org.fossasia.openevent.general.event
+package org.fossasia.openevent.general.ticket
 
 import android.arch.lifecycle.Observer
 import android.os.Bundle
