+++ app/src/main/java/org/fossasia/openevent/general/ticket/TicketsViewModel.kt
@@ -1,4 +1,4 @@
-package org.fossasia.openevent.general.event
+package org.fossasia.openevent.general.ticket
 
 import android.arch.lifecycle.MutableLiveData
 import android.arch.lifecycle.ViewModel
